+++ hata/ext/slash/utils.py
@@ -2,6 +2,7 @@ __all__ = ()
 
 from ...discord.client import Client
 
+
 UNLOADING_BEHAVIOUR_DELETE = 0
 UNLOADING_BEHAVIOUR_KEEP = 1
 UNLOADING_BEHAVIOUR_INHERIT = 2
@@ -75,3 +76,5 @@ def runtime_sync_hook_is_client_running(client):
     return client.running
 
 RUNTIME_SYNC_HOOKS.append(runtime_sync_hook_is_client_running)
+
+